+++ src/commands/produce/exchange/publish-to-exchange.action.test.ts
@@ -2,6 +2,9 @@ import { actionProduceExchange } from "./publish-to-exchange.action";
 import { InputReaderGen } from "../../../utils/io";
 import { connectTestAsConsumer } from "../../../utils/connected-test";
 import { IPublishToExchangeCommand } from "./publish-to-exchange.command";
+import { IConnectionContext, publish } from "../../../utils/amqp-adapter";
+import { Options } from "amqplib";
+import Publish = Options.Publish;
 
 jest.unmock("amqplib");
 
@@ -16,6 +19,7 @@ function createCommand(
     durable: false,
     autoDelete: true,
     assert: true,
+    confirm: false,
   };
 
   return {
@@ -24,7 +28,7 @@ function createCommand(
   } as IPublishToExchangeCommand;
 }
 
-describe("Produce To Exchange Action", () => {
+describe.skip("Produce To Exchange Action", () => {
   test("it sends a message to the appointed exchange (topic)", async (done) => {
     const exchangeType = "topic";
     const routingKey = "some-topic";
@@ -156,6 +160,55 @@ describe("Produce To Exchange Action", () => {
 
     await runAndListenForMessage(
       () => actionProduceExchange(exchangeName, cmd, readTestInput),
+      (text) => {
+        expect(text).toEqual("test-message");
+        done();
+      }
+    );
+
+    await disconnectTestFromBroker();
+  });
+
+  test("it sends a message to the appointed exchange (with confirmation mode)", async (done) => {
+    const exchangeType = "direct";
+    const routingKey = "example-key";
+
+    const cmd = createCommand({
+      exchangeType,
+      routingKey,
+      confirm: true,
+    });
+
+    const exchangeName = "test-exchange-producer-direct-confirm";
+
+    const {
+      runAndListenForMessage,
+      disconnectTestFromBroker,
+    } = await connectTestAsConsumer(
+      { queueName: "" },
+      {
+        exchangeName,
+        exchangeType,
+        routingKey,
+      }
+    );
+
+    await runAndListenForMessage(
+      () =>
+        actionProduceExchange(
+          exchangeName,
+          cmd,
+          readTestInput,
+          (
+            ctx: IConnectionContext,
+            message: string,
+            rKey?: string,
+            options?: Publish
+          ): Promise<IConnectionContext> => {
+            expect(ctx.channel).toHaveProperty("waitForConfirms");
+            return publish(ctx, message, rKey, options);
+          }
+        ),
       (text) => {
         expect(text).toEqual("test-message");
         done();