+++ src/commands/produce/queue/send-to-queue.command.ts
@@ -6,6 +6,7 @@ export interface ISendToQueueCommand extends Command {
   assert: boolean;
   durable: boolean;
   autoDelete: boolean;
+  confirm: boolean;
 }
 
 export function buildSendToQueueCommand(): ISendToQueueCommand {
@@ -29,5 +30,10 @@ export function buildSendToQueueCommand(): ISendToQueueCommand {
       false
     )
     .option("--autoDelete", "Marks the used queue for automatic deletion", true)
+    .option(
+      "--confirm",
+      "Use publisher confirms to wait for the broker to confirm if the message was handled",
+      false
+    )
     .action(actionProduceQueue) as ISendToQueueCommand;
 }